Rewrote the "find hard disks on windows" to use a dll instead. The reasons:
A) FSUTIL is internationalized, so on non-english installs, we can't figure out if a disk is fixed.

B) I trust this better than fsutil for user control access (a.k.a. popups from hell) on vista.

C) fsutil.exe is in C:\windows\system32 so pretty sure it would be there, but there's always a chance it's not in the PATH, or otherwise not accessible.