[suspect:eclipse] Fix for issue 633: refactor script 'extract local variable' did not work inside for example SneakyThrows-protected methods.
This fix is reverting the change in e835daa9.

During conversion of the internal.ast to the dom AST, a flag is introduced
named 'ORIGINAL'. Commit e835daa9 always explicitly cleared it for
all lombok-generated nodes, in an attempt to try and do the appropriate thing,
treating the ORIGINAL flag as a simile for 'generated code'.

Unfortunately, for eclipse code analysis, all our generated code should be
treated as if it's "really" there, and in general this doesn't work because
eclipse's concept of 'generated' applies to explicit super() calls and the like,
not the kind of stuff lombok introduces.