Don't modify "from" javadoc in copyJavadoc
This, together with the parent commit, means that multiple generated
members can now "extract" the same elements from the javadoc on a field.
For example, a `@Setter` and a `@Wither` can now both receive a `@param`
tag from the javadoc on their defining field, whereas previously the
first one to be processed would remove the tag from the field so it
didn't show up in the javadoc on the other.

This fixes the bug exposed by the test added in commit
480129abf69fd6c1a1233fd0f16fd39fa7ed9edb.