[tests] removed a ridiculously exotic feature that still works, but is too hard to test
Specifically, using as custom logger topic a concatenated string, like literally `"A" + "B"`. It works,
but depending on javac and phase of the moon that ends up as `"A" + "B"` or `"AB"`, and I don't
think it's worthwhile for lombok to try to be consistent in this and test for that.