[records] [@NonNull] slight refactor to ensure any generated canonical constructor is visible to downstream APs
We used to remove the canonical constructor on a record that javac generates, and replace it with our own, as we put null checks in it.
However, the type mirrors and such Annotation Processors use already have the implicit canonical constructor available. Instead of
trying to remove those, let's just take the existing implicit constructor and add to that (and mark it explicit, of course).