issue #3517
modified annotation to have shorter name for prefix, suffix, added option to camelcase the variable name and the method name. Created util to extract the var name and met name from the requested name and the annotation.
Changed the tests to have the var match the getter, and implemented the var name  and assignment replacement in the node.